openHandles added , bug fixes
